--- a/kit/datetime/datetime_test.go
+++ b/kit/datetime/datetime_test.go
@@ -5,24 +5,23 @@
 package datetime_test
 
 import (
+	"fmt"
 	"testing"
 	"time"
 
 	"github.com/muktihari/fit/kit/datetime"
 	"github.com/muktihari/fit/profile/basetype"
-	"github.com/muktihari/fit/profile/typedef"
-	"github.com/muktihari/fit/proto"
 )
 
 func TestToTime(t *testing.T) {
 	tt := []struct {
 		name string
-		u32  any
+		u32  uint32
 		time time.Time
 	}{
 		{
 			name: "Thu, 30 Dec 2021 21:52:08 GMT",
-			u32:  proto.Uint32(uint32(1009835528)),
+			u32:  uint32(1009835528),
 			time: time.Date(2021, time.December, 30, 21, 52, 8, 00, time.UTC),
 		},
 		{
@@ -32,22 +31,17 @@ func TestToTime(t *testing.T) {
 		},
 		{
 			name: "Thu, 30 Dec 2021 21:52:08 GMT",
-			u32:  typedef.DateTime(1009835528),
+			u32:  uint32(1009835528),
 			time: time.Date(2021, time.December, 30, 21, 52, 8, 00, time.UTC),
 		},
 		{
 			name: "Thu, 30 Dec 2021 21:52:08 GMT",
-			u32:  typedef.LocalDateTime(1009835528),
+			u32:  uint32(1009835528),
 			time: time.Date(2021, time.December, 30, 21, 52, 8, 00, time.UTC),
 		},
 		{
 			name: "nil",
-			u32:  nil,
-			time: time.Time{},
-		},
-		{
-			name: "struct{}{}",
-			u32:  struct{}{},
+			u32:  basetype.Uint32Invalid,
 			time: time.Time{},
 		},
 	}
@@ -70,17 +64,12 @@ func TestToLocalTime(t *testing.T) {
 
 	tt := []struct {
 		name   string
-		value  any
+		value  time.Time
 		result time.Time
 	}{
 		{
-			name:   "utc to utc+7",
-			value:  uint32(1029622579), // Tue, 16 Aug 2022 22:16:19 GMT
-			result: time.Date(2022, 8, 17, 05, 16, 19, 0, locJakarta),
-		},
-		{
 			name:   "unsupported tipe",
-			value:  1029622579, // int int
+			value:  time.Time{},
 			result: time.Time{},
 		},
 		{
@@ -90,8 +79,8 @@ func TestToLocalTime(t *testing.T) {
 		},
 	}
 
-	for _, tc := range tt {
-		t.Run(tc.name, func(t *testing.T) {
+	for i, tc := range tt {
+		t.Run(fmt.Sprintf("[%d] %s", i, tc.name), func(t *testing.T) {
 			_time := datetime.ToLocalTime(tc.value, 7)
 			if _time.Format(time.RFC3339) != tc.result.Format(time.RFC3339) { // can't use time equal since the loc object is different
 				t.Fatalf("expected: %s, got: %s", tc.result.Format(time.RFC3339), _time.Format(time.RFC3339))
@@ -200,5 +189,19 @@ func TestEpoch(t *testing.T) {
 	expected := time.Date(1989, time.December, 31, 0, 0, 0, 0, time.UTC)
 	if !ep.Equal(expected) {
 		t.Fatalf("expected: %s, got: %s", expected, ep)
+	}
+}
+
+func BenchmarkToTime(b *testing.B) {
+	for i := 0; i < b.N; i++ {
+		_ = datetime.ToTime(uint32(1009835528))
+	}
+}
+
+func BenchmarkToLocalTime(b *testing.B) {
+	t := datetime.ToTime(uint32(1009835528))
+	b.ResetTimer()
+	for i := 0; i < b.N; i++ {
+		datetime.ToLocalTime(t, 8)
 	}
 }