Breaking Changes:
- ToTime now accepts uint32 instead of any
- ToLocalTime now accepts time.Time instead on any and it is no longer allocate.

Since we use proto.Value, all timestamp will be represented as uint32 and in mesgdef's structures all timestamp will be represented as time.Time. There are no timestamp being represented as typedef.DateTime or typedef.LocalDateTime anymore. This way we can no longer need to convert the value to an interface, this will improve performance a little bit and we have more clear context when working with time.

Benchmark:
```js
goos: darwin
goarch: amd64
pkg: github.com/muktihari/fit/kit/datetime
cpu: Intel(R) Core(TM) i5-5257U CPU @ 2.70GHz
              │   old.txt    │               new.txt                │
              │    sec/op    │    sec/op     vs base                │
ToTime-4        10.870n ± 1%   7.452n ±  0%  -31.44% (p=0.000 n=10)
ToLocalTime-4   193.70n ± 1%   10.68n ± 23%  -94.48% (p=0.000 n=10)
geomean          45.89n        8.923n        -80.55%

              │   old.txt    │                 new.txt                 │
              │     B/op     │    B/op     vs base                     │
ToTime-4        0.000 ± 0%     0.000 ± 0%         ~ (p=1.000 n=10) ¹
ToLocalTime-4   168.0 ± 0%       0.0 ± 0%  -100.00% (p=0.000 n=10)
geomean                    ²               ?                       ² ³
¹ all samples are equal
² summaries must be >0 to compute geomean
³ ratios must be >0 to compute geomean

              │   old.txt    │                 new.txt                 │
              │  allocs/op   │ allocs/op   vs base                     │
ToTime-4        0.000 ± 0%     0.000 ± 0%         ~ (p=1.000 n=10) ¹
ToLocalTime-4   4.000 ± 0%     0.000 ± 0%  -100.00% (p=0.000 n=10)
geomean                    ²               ?                       ² ³
¹ all samples are equal
² summaries must be >0 to compute geomean
³ ratios must be >0 to compute geomean

```